feat: add hidden dumb-llb flag (#225)
without this, you have to manually enable the bool to get a dumped llb.

I don't think it makes sense to enable this as a public flag at any point soon (or ever)
but for development having this available as a CLI flag will make life easier and
making it a hidden flag is more self-documenting to developers working with the project.